pyproject.toml
[tool.poetry]
name = "api-cartoes"
version = "0.1.0"
description = ""
authors = ["lazzarusxd"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
pika = "^1.3.2"
cryptography = "^43.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.9.2"}
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
pytz = "^2024.2"
httpx = "^0.27.2"
python-multipart = "^0.0.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
